+++ tests/test_tile.cpp
@@ -7,7 +7,7 @@ TEST(TileTest, AllZeroTileShouldBeAllWhite)
 	Tile tile(data);
 	const Tile::TileRGBArray& pixels = tile.toRGB();
 
-	for (int i = 0; i < pixels.size(); i++)
+	for (size_t i = 0; i < pixels.size(); i++)
 	{
 		ASSERT_EQ(pixels[i], Tile::COLOR_WHITE);
 	}
@@ -20,7 +20,7 @@ TEST(TileTest, FilledTileShouldBeAllBlack)
 	Tile tile(data);
 	const Tile::TileRGBArray& pixels = tile.toRGB();
 
-	for (int i = 0; i < pixels.size(); i++)
+	for (size_t i = 0; i < pixels.size(); i++)
 	{
 		ASSERT_EQ(pixels[i], Tile::COLOR_BLACK);
 	}
@@ -29,7 +29,7 @@ TEST(TileTest, FilledTileShouldBeAllBlack)
 TEST(TileTest, LSBTileShouldBeAllLightGray)
 {
 	Tile::TileDataArray data = {};
-	for (int i = 0; i < data.size(); ++i)
+	for (size_t i = 0; i < data.size(); ++i)
 	{
 		if (i % 2 == 0)
 		{
@@ -43,7 +43,7 @@ TEST(TileTest, LSBTileShouldBeAllLightGray)
 	Tile tile(data);
 	const Tile::TileRGBArray& pixels = tile.toRGB();
 
-	for (int i = 0; i < pixels.size(); i++)
+	for (size_t i = 0; i < pixels.size(); i++)
 	{
 		ASSERT_EQ(pixels[i], Tile::COLOR_LIGHT_GRAY);
 	}
@@ -52,7 +52,7 @@ TEST(TileTest, LSBTileShouldBeAllLightGray)
 TEST(TileTest, MSBTileShouldBeAllDarkGray)
 {
 	Tile::TileDataArray data = {};
-	for (int i = 0; i < data.size(); ++i)
+	for (size_t i = 0; i < data.size(); ++i)
 	{
 		if (i % 2 == 1)
 		{
@@ -66,7 +66,7 @@ TEST(TileTest, MSBTileShouldBeAllDarkGray)
 	Tile tile(data);
 	const Tile::TileRGBArray& pixels = tile.toRGB();
 
-	for (int i = 0; i < pixels.size(); i++)
+	for (size_t i = 0; i < pixels.size(); i++)
 	{
 		ASSERT_EQ(pixels[i], Tile::COLOR_DARK_GRAY);
 	}
